Delete TRACSStructuredCodeAnnotation.md
Card working but much longer than the rest which pulled most of the other cards elements into the centre of the card space. Need to change styling so that each element of a card aligns to the top rather than the middle of its enclosing parent element. May need to consider the ordering of examples by size rather than alphabetical order.